source/requirements.txt
openai
pydub
python-dotenv